+++ 制作/Task_EnemySkeleton.cpp
@@ -484,7 +484,7 @@ namespace  EnemySkeleton
 
 		this->unHitTime = 10;
 		this->hp.Addval(-at_.power);
-		ge->TotalDamage += at_.power;
+		ge->TotalDamage += static_cast<int>(at_.power);
 		if (this->hp.vnow <= 0) {
 			this->UpdateMotion(Motion::Lose);
 			ge->TotalEnemyKill += 1;
